Refactor main.js to use ES module syntax and improve organization; add configuration, diagram operations, file operations, and utility modules for better code structure and maintainability.
