+++ demoSite/index.html
@@ -15,7 +15,7 @@
     <script src="js/config-ui.js" defer></script>
     <script src="js/ai-assistant.js" defer></script>
     <script src="js/code-history.js" defer></script>
-    <script src="js/main.js" defer></script>
+    <script src="js/main.js" type="module"></script>
 
 </head>
